feat(nx-deploy-it): aws - angular universal
- add support for angular universal on aws
