+++ .vscode/settings.json
@@ -13,7 +13,9 @@
     "titleBar.inactiveForeground": "#15202b99",
     "statusBar.background": "#42b883",
     "statusBarItem.hoverBackground": "#359268",
-    "statusBar.foreground": "#15202b"
+    "statusBar.foreground": "#15202b",
+    "statusBar.border": "#42b883",
+    "titleBar.border": "#42b883"
   },
   "peacock.color": "#42b883",
   "typescript.tsdk": "node_modules/typescript/lib",
