+++ libs/nx-deploy-it/src/adapter/angular-universal/deployment-type.enum.ts
@@ -1,0 +1,4 @@
+export enum ANGULAR_UNIVERSAL_DEPLOYMENT_TYPE {
+  PRERENDERING = 'prerendering',
+  SERVER_SIDE_RENDERING = 'ssr'
+}
