+++ libs/nx-deploy-it/src/adapter/base.adapter.model.ts
@@ -1,0 +1,7 @@
+import { PROVIDER } from '../utils/provider';
+import { JsonObject } from '@angular-devkit/core';
+
+export interface NxDeployItBaseOptions extends JsonObject {
+  provider: PROVIDER;
+  project: string;
+}
